Explict Node bump version, fix argument parsing.

+++ lib/dots/bin/node/release.zsh
@@ -10,24 +10,48 @@ source $dots <<- usage
     Bump version number and publish a release.
 usage
 
-zparseopts -D -- -help=usage h=usage \
-                 -title:=title i:=title \
-                 -prefix:=prefix p:=prefix
+zparseopts -a opts -D -- -help h -title: t: -prefix: p: -version: v: d -dry-run
+
+index=1
+while [ $index -le $#opts ]; do
+    case "${opts[$index]}" in
+        -h|--help)
+            echo help
+            ;;
+        -d|--dry-run)
+            dry_run=1
+            ;;
+        -t|--title)
+            let index+=1
+            title=$opts[$index]
+            ;;
+        -p|--prefix)
+            let index+=1
+            prefix=$opts[$index]
+            ;;
+        -v|--version)
+            let index+=1
+            bump=$opts[$index]
+            ;;
+    esac
+    let index+=1
+done
 
 version=$(dots node package version)
-if [ "$1" = "major" ]; then
-    major=${version%%.*}
-    bump=$(( $major + 1 )).0.0
-elif [ "$1" = "minor" ]; then
-    major=${version%%.*}
-    minor=${version#*.}
-    minor=${minor%.*}
-    echo $minor
-    bump=$major.$(( $minor + 1 )).0
-else
-    majmin=${version%.*}
-    micro=${version##*.}
-    bump=$majmin.$(( $micro + 1 ))
+if [ -z "$bump" ]; then
+    if [ "$1" = "major" ]; then
+        major=${version%%.*}
+        bump=$(( $major + 1 )).0.0
+    elif [ "$1" = "minor" ]; then
+        major=${version%%.*}
+        minor=${version#*.}
+        minor=${minor%.*}
+        bump=$major.$(( $minor + 1 )).0
+    else
+        majmin=${version%.*}
+        micro=${version##*.}
+        bump=$majmin.$(( $micro + 1 ))
+    fi
 fi
 
 if [ -z "$prefix" ]; then
@@ -48,6 +72,7 @@ if [ -z "$title" ]; then
 fi
 
 echo "$title $prefix$version -> $prefix$bump"
+[ "$dry_run" -eq 1 ] && exit
 sed 's/\("version":.*"\)'$version'/\1'$bump'/' package.json > package.json.tmp
 mv package.json.tmp package.json
 git add .
